requirements.txt
# Core runtime dependencies
pandas>=1.5.0
python-dotenv>=0.19.0
requests>=2.28.0
schedule>=1.1.0

# Chart generation is only needed if you call BusinessMetrics.generate_summary_html().
plotly>=5.5.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
